Remove DevKitARM and proprietary dependencies
This fork included the DefaultArm7 and Loader folders which cannot be compiled using the current DevKitARM releases. They are not needed for building the bootloader.nds needed by HiyaCFW so they were removed.

Also, the ndstool.cpp file included references to a couple of private functions (EncryptSecureArea and EnDecryptSecureArea) that were not available so their calls were removed.